backend/requirements.txt
alembic==1.15.1
annotated-types==0.7.0
anyio==4.8.0
assistant-stream==0.0.5
asyncpg==0.30.0
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
duckduckgo-search==7.5.1
email-validator==2.2.0
eval-type-backport==0.2.2
executing==2.2.0
fastapi==0.115.11
fastapi-sqlalchemy==0.2.1
googleapis-common-protos==1.69.1
greenlet==3.1.1
griffe==1.6.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib-metadata==8.5.0
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jwcrypto==1.5.6
langchain-core==0.3.43
langchain-openai==0.2.14
langgraph==0.2.76
langgraph-checkpoint==2.0.18
langgraph-sdk==0.1.55
langsmith==0.3.13
logfire==3.7.1
logfire-api==3.7.1
lxml==5.3.1
mako==1.3.9
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
msgpack==1.1.0
openai==1.65.4
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
orjson==3.10.15
packaging==24.2
pip==23.3.1
primp==0.14.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic-ai-slim==0.0.36
pydantic-core==2.27.2
pydantic-graph==0.0.36
pydantic-settings==2.8.1
pygments==2.19.1
pyjwt==2.10.1
python-dotenv==1.0.1
python-multipart==0.0.20
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
setuptools==69.0.2
sniffio==1.3.1
sqlalchemy==2.0.39
starlette==0.46.1
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typing-extensions==4.12.2
typing-inspection==0.4.0
urllib3==2.3.0
uvicorn==0.23.2
wrapt==1.17.2
zipp==3.21.0
zstandard==0.23.0
